requirements.txt
pyocr==0.7.2
python-xlib==0.27
Pillow==7.1.1
python-daemon==2.2.4
